reorganize `key`
- Move the unencrypted and encrypted implementations of `Key` into the `key` module.
