+++ src/key/unencrypted.rs
@@ -1,12 +1,11 @@
-//! [`Key`] management for Anthropic API keys.
-
-// TODO: Remove this dependency for wasm32 and find an alternative. It's not a
-// super idea to use this in a web app but wasm32 also has server use cases.
-// This is the only thing that prevents this from building on wasm32.
-use memsecurity::zeroize::Zeroizing;
+//! Unencrypted [`Key`] management for Anthropic API keys.
+use zeroize::{ZeroizeOnDrop, Zeroizing};
 
 /// The length of an Anthropic API key in bytes.
 pub const LEN: usize = 108;
+
+/// Type alias for an Anthropic API key.
+type Arr = [u8; LEN];
 
 /// Error for when a key is not 108 bytes.
 #[derive(Debug, thiserror::Error)]
@@ -16,25 +15,24 @@ pub struct InvalidKeyLength {
     pub actual: usize,
 }
 
-/// Stores an Anthropic API key securely. The API key is encrypted in memory.
-/// The object features a [`Display`] implementation that can be used to write
-/// out the key. **Be sure to zeroize whatever you write it to**. Prefer
-/// [`Key::read`] if you want a return value that will automatically zeroize
-/// the key on drop.
+/// Stores an Anthropic API key securely. The object features a [`Display`]
+/// implementation that can be used to write out the key. **Be sure to zeroize
+/// whatever you write it to**. The key is zeroized on drop.
 ///
 /// [`Display`]: std::fmt::Display
-#[derive(Debug)]
+#[derive(Debug, ZeroizeOnDrop)]
 pub struct Key {
-    // FIXME: `memsecurity` does not build on wasm32. Find a solution for web.
-    // The `keyring` crate may work, but I'm likewise not sure if it builds on
-    // wasm32. It's not listed in the platforms so likely not.
-    mem: memsecurity::EncryptedMem,
+    mem: Arr,
 }
 
 impl Key {
     /// Read the key. The key is zeroized on drop.
-    pub fn read(&self) -> memsecurity::ZeroizeBytes {
-        self.mem.decrypt().unwrap()
+    // We can't return a &str becuase the other implementation of Key::read
+    // returns a memsecurity::ZeroizeBytes, and we can't return a reference to
+    // that because it's a temporary value, so this returns a slice instead,
+    // which has more or less the same public API.
+    pub fn read(&self) -> &[u8] {
+        &self.mem
     }
 }
 
@@ -44,21 +42,15 @@ impl TryFrom<String> for Key {
     /// Create a new key from a string securely. The string is zeroized after
     /// conversion.
     fn try_from(s: String) -> Result<Self, Self::Error> {
-        // This just unwraps the internal Vec<u8> so the data can still be
-        // zeroized.
         let v = Zeroizing::new(s.into_bytes());
+        let mut arr: Arr = [0; LEN];
         if v.len() != LEN {
             let actual = v.len();
             return Err(InvalidKeyLength { actual });
         }
 
-        let mut mem = memsecurity::EncryptedMem::new();
-
-        // Unwrap is desirable here because this can only fail if encryption
-        // is broken, which is a catastrophic failure.
-        mem.encrypt(&v).unwrap();
-
-        Ok(Self { mem })
+        arr.copy_from_slice(&v);
+        Ok(Key { mem: arr })
     }
 }
 
@@ -69,11 +61,9 @@ impl std::fmt::Display for Key {
     /// Prefer [`Self::read`] if you want a return value that will automatically
     /// zeroize the key on drop.
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // Zeroized on drop
-        let key = self.read();
-        // Unwrap can never panic because a Key can only be created from a
-        // String which is guaranteed to be valid UTF-8.
-        let key_str = std::str::from_utf8(key.as_ref()).unwrap();
+        // Unwrap can never panic because a Key can only be created from String
+        // whic is guaranteed to be valid UTF-8.
+        let key_str = std::str::from_utf8(self.read()).unwrap();
         write!(f, "{}", key_str)
     }
 }
